+++ packages/react/src/components/DesignLibrary/models.ts
@@ -1,4 +1,8 @@
-import { ImportEntry, ImportEntryInfo } from '@sitecore-content-sdk/content/codegen';
+import {
+  GeneratedComponentData,
+  ImportEntry,
+  ImportEntryInfo,
+} from '@sitecore-content-sdk/content/codegen';
 import {
   ComponentFields,
   ComponentParams,
@@ -7,7 +11,6 @@ import {
 } from '@sitecore-content-sdk/content/layout';
 import { DesignLibraryStatus } from '@sitecore-content-sdk/content/editing';
 import { Page } from '@sitecore-content-sdk/content/client';
-import { ComponentPreviewEventArgs } from '@sitecore-content-sdk/content/codegen';
 import { ComponentMap } from '../sharedTypes';
 
 export type ImportMapImport = {
@@ -62,11 +65,14 @@ export type DesignLibraryVariantGenerationEventsProps = DesignLibraryPreviewEven
    */
   importMap?: ImportEntryInfo[];
   /**
-   * Any error that occurred while loading the import map to be posted as a message to the Design Studio.
+   * Any error that occurred during initialization of the component:
+   * - importMap error
+   * - error fetching the generated component data from secured endpoint
+   * - error during generation of the component on the server side
    */
-  importMapError?: string;
+  componentInitError?: string;
   /**
-   * The preview component data received from design library.
+   * The generated component data received from design library.
    */
-  previewComponentData?: ComponentPreviewEventArgs;
+  generatedComponentData?: GeneratedComponentData;
 };